Cadastro_exames/requirements.txt.txt
tkcalendar
mysql-connector-python
python-dotenv